Add All Records

+++ Attendance.Domain/Models/LeaveBalance.cs
@@ -27,9 +27,9 @@ namespace Attendance.Domain.Models
         [ForeignKey("LeaveMasterId")]
         public int? LeaveMasterId { get; set; }
         public LeaveMasterDto? LeaveMaster { get; set; }
-        public int AssignedLeaves { get; set; }
-        public int UsedLeaves { get; set; }
-        public int RemainingLeaves { get; set; }
-        public int? ExtraLeaves { get; set; }
+        public double AssignedLeaves { get; set; }
+        public double UsedLeaves { get; set; }
+        public double RemainingLeaves { get; set; }
+        public double? ExtraLeaves { get; set; }
 	}
 }
